requirements.txt
fastapi
uvicorn
python-multipart
openai
langchain
langchain-openai
chromadb
sentence-transformers
pydantic
pydantic-settings
python-dotenv
soundfile
numpy
requests
aiofiles
websockets
pydub
transformers
torch
accelerate 